# model-pricing-backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "大模型中转平台：模型数据中心 + 进货/出货两条定价线。统一响应包 {code,message,data,requestId}；金额一律字符串传输；时间 ISO 8601 带时区。\n错误码：400 参数校验、401 未认证、403 无权限、409 幂等/状态冲突、423 主体冻结、429 限流。\n写操作（产生业务新版本/影响资金合同）必须携带 Idempotency-Key 头。\n分页：?page=1\u0026size=20 或 ?cursor=\u0026size=，统一返回 {list,total,page,size}。",
        "title": "模型管理与定价中心 API",
        "contact": {
            "name": "后端"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/internal/model-applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "模型运营查看全部供应商的申请，支持按状态与供应商筛选。dup_top3 为提交时的查重候选，供审核参考。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型申请"
                ],
                "summary": "模型申请列表（内部）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态筛选（空=全部）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "供应商 ID（空=全部）",
                        "name": "supplier_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小（默认 20，上限 200）",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data={list,total,page,size}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "403": {
                        "description": "code=10003 无 M1:V 权限",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/internal/model-applications/{id}/decision": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "APPROVE / MERGE 必须指定 target_sku_id（指向已存在的 model_sku）；REJECT 必须填 reason。仅 SUBMITTED 可审核，终态再次审核返回 10005。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模型申请"
                ],
                "summary": "审核模型申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "model_application.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "{action, target_sku_id?, reason?}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/supplier.ApplicationDecisionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=ModelApplication（status 已推进）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法（缺 target_sku_id / reason）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "403": {
                        "description": "code=10003 无 M1:E 权限",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 申请或目标 SKU 不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "409": {
                        "description": "code=10005 申请已终结（不可重复审核）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/supplier/model-applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "行级过滤：仅返回本供应商主体的申请。支持 status 筛选（SUBMITTED/MERGED/APPROVED/REJECTED）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商-新模型申请"
                ],
                "summary": "我的模型申请列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态筛选（空=全部）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小（默认 20，上限 200）",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data={list,total,page,size}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "供应商提交拟引入的新模型。payload 为自由结构（能力/上下文/模态/期望定价等）。提交时按 model_name 与已有 SKU 做相似度查重，结果落在 dup_top3 供内部审核参考（只提示不拦截）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "供应商-新模型申请"
                ],
                "summary": "提交新模型申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "{model_name, vendor_id?, payload}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/supplier.SubmitApplicationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=ModelApplication（status=SUBMITTED）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法（model_name/payload 缺失或非 JSON 对象）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "403": {
                        "description": "code=10003 非供应商主体",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0=成功",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "业务数据",
                    "type": "object"
                },
                "message": {
                    "description": "错误消息，成功时 \"ok\"",
                    "type": "string",
                    "example": "ok"
                },
                "requestId": {
                    "description": "请求追踪号（X-Request-Id）",
                    "type": "string",
                    "format": "uuid",
                    "example": "01929d81-..."
                }
            }
        },
...
        "supplier.ApplicationDecisionInput": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "APPROVE / MERGE / REJECT",
                    "type": "string"
                },
                "reason": {
                    "description": "REJECT 时必填",
                    "type": "string"
                },
                "target_sku_id": {
                    "description": "TargetSKUID APPROVE / MERGE 时必填：指向已存在的 model_sku。\n  APPROVE → 该 SKU 是为本申请新建/对应的 SKU\n  MERGE   → 被合并到的已有 SKU",
                    "type": "integer"
                }
            }
        },
...
        "supplier.SubmitApplicationInput": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003ctoken\u003e，登录后获得",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}